+++ engine/server/main_server_linux.cpp
@@ -82,7 +82,7 @@ int main(int argc, char** argv)
 	logSettings.console = true;
 	logSettings.flushAlways = true;
 	logSettings.filePath = config.GetString("log.file", "engine.log");
-	logSettings.rotation_size_mb = static_cast<size_t>(std::max(0, config.GetInt("log.rotation_size_mb", 10)));
+	logSettings.rotation_size_mb = static_cast<size_t>(std::max(static_cast<int64_t>(0), config.GetInt("log.rotation_size_mb", 10)));
 	logSettings.retention_days = static_cast<int>(config.GetInt("log.retention_days", 7));
 	engine::core::Log::Init(logSettings);
 
@@ -145,7 +145,7 @@ int main(int argc, char** argv)
 
 	engine::server::SecurityAuditLog auditLog;
 	std::string auditPath = config.GetString("security.audit_log_path", "security_audit.log");
-	size_t rotationMb = static_cast<size_t>(std::max(0, config.GetInt("log.rotation_size_mb", 10)));
+	size_t rotationMb = static_cast<size_t>(std::max(static_cast<int64_t>(0), config.GetInt("log.rotation_size_mb", 10)));
 	int retentionDays = static_cast<int>(config.GetInt("log.retention_days", 7));
 	if (auditLog.Init(auditPath, rotationMb, retentionDays))
 		LOG_INFO(Net, "[ServerMain] SecurityAuditLog opened: {}", auditPath);
